backend/requirements.txt
fastapi>=0.115.0
uvicorn[standard]>=0.32.0
sqlalchemy>=2.0.36
pydantic>=2.10.0
python-multipart>=0.0.12
python-dateutil>=2.9.0
mt940==0.7.0
pydantic-settings>=2.6.0
psycopg2-binary>=2.9.0
python-jose[cryptography]>=3.3.0
bcrypt>=4.0.0
python-dotenv>=1.0.0
